requirements.txt
streamlit
scikit-learn
streamlit_extras
nltk
unidecode
tqdm
wordcloud
requests
pandas
joblib
xgboost
afinn
emoji
emosent-py